CLI11
CI Fix

Remove copyright check from github config folder

Clean up

Rename from_raw -> convert_to

Update Copyright

Improve help menu

Comment Command classes

Remove Docopt

More comments

Remove docopt from cmake

clang-format

Explicitely state how many values a flag should expect

+++ framework/core/descriptor_set.cpp
@@ -1,4 +1,4 @@
-/* Copyright (c) 2019-2020, Arm Limited and Contributors
+/* Copyright (c) 2019-2021, Arm Limited and Contributors
  *
  * SPDX-License-Identifier: Apache-2.0
  *
@@ -83,7 +83,7 @@ void DescriptorSet::prepare()
 				size_t uniform_buffer_range_limit = device.get_gpu().get_properties().limits.maxUniformBufferRange;
 				size_t storage_buffer_range_limit = device.get_gpu().get_properties().limits.maxStorageBufferRange;
 
-				size_t buffer_range_limit = buffer_info.range;
+				size_t buffer_range_limit = static_cast<size_t>(buffer_info.range);
 
 				if ((binding_info->descriptorType == VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER || binding_info->descriptorType == VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC) && buffer_range_limit > uniform_buffer_range_limit)
 				{
